pyproject.toml
[tool.poetry]
name = "artistic"
version = "0.1.0"
description = ""
authors = ["Trevor Jones <trevordavidjones@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flake8 = "^3.9.0"
isort = "^5.8.0"
Flask = "^1.1.2"
python-dotenv = "^0.16.0"
Flask-Migrate = "^2.7.0"
db = "^0.1.1"
init = "^0.1.0"
Flask-Script = "^2.0.6"
psycopg2 = "^2.8.6"
google-cloud-storage = "^1.37.0"
google-cloud-notebooks = "^0.1.2"
Flask-Login = "^0.5.0"
passlib = "^1.7.4"
Pillow = "^8.2.0"
grpcio = "^1.37.0"
Flask-Mail = "^0.9.1"
opencv-python = "^4.5.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
